Refactoring app specific locator builders
