+++ packages/selenium-ide/src/content/locatorBuilders_custom.js
@@ -28,6 +28,7 @@ window.LocatorBuilders = LocatorBuilders
 LocatorBuilders.prototype.detach = function() {}
 
 LocatorBuilders.prototype.buildWith = function(name, e, opt_contextNode) {
+  console.log('buildWith: ' + name);
   return LocatorBuilders.builderMap[name].call(this, e, opt_contextNode)
 }
 
@@ -61,6 +62,7 @@ LocatorBuilders.prototype.buildAll = function(el, ignoreInnerText) {
     if (elClass && elClass.indexOf('slds-file-selector__button') > -1)
       return [];
   }
+  console.log('before displayname: ' + LocatorBuilders.order.length);
   this.displayName = this.getDisplayName(e, ignoreInnerText);
   let locator
   let locators = []
@@ -87,6 +89,7 @@ LocatorBuilders.prototype.buildAll = function(el, ignoreInnerText) {
       }
     } catch (e) {
       // TODO ignore the buggy locator builder for now
+      console.log('locator exception: ' + e);
       //this.log.debug("locator exception: " + e);
       //  this.logging(finderName + ":locator Exception: " + e)
     }
@@ -111,6 +114,42 @@ LocatorBuilders.prototype.findElement = function(loc) {
  * Class methods
  */
 
+LocatorBuilders.prototype.buildReactTableRowData = function(e) {
+  return undefined;
+};
+
+LocatorBuilders.prototype.cleanup = function() {
+  console.log('cleanup step')
+  LocatorBuilders.order = []
+  LocatorBuilders.builderMap = {}
+  LocatorBuilders._preferredOrder = []
+  LocatorBuilders.PREFERRED_ATTRIBUTES = [
+    'id',
+    'name',
+    'value',
+    'type',
+    'action',
+    'onclick'
+  ]
+  debugger;
+  if (!LocatorBuilders.getDisplayNameFn)
+    LocatorBuilders.getDisplayNameFn = this.getDisplayName;
+  else
+    this.getDisplayName = LocatorBuilders.getDisplayNameFn;
+  this.buildReactTableRowData = function(e) {
+    return undefined;
+  };
+}
+
+LocatorBuilders.PREFERRED_ATTRIBUTES = [
+  'id',
+  'name',
+  'value',
+  'type',
+  'action',
+  'onclick'
+]
+
 LocatorBuilders.order = []
 LocatorBuilders.builderMap = {}
 LocatorBuilders._preferredOrder = []
@@ -118,6 +157,7 @@ LocatorBuilders._preferredOrder = []
 // classObservable(LocatorBuilders);
 
 LocatorBuilders.add = function(name, finder) {
+  console.log('adding finderName: ' + name);
   this.order.push(name)
   this.builderMap[name] = finder
   this._orderChanged()
@@ -290,137 +330,9 @@ LocatorBuilders.prototype.preciseXPath = function(xpath, e) {
 // order listed dictates priority
 // e.g., 1st listed is top priority
 
-LocatorBuilders.add('css:data-attr', function cssDataAttr(e) {
-  const dataAttributes = ['data-test', 'data-test-id']
-  for (let i = 0; i < dataAttributes.length; i++) {
-    const attr = dataAttributes[i]
-    const value = e.getAttribute(attr)
-    if (attr) {
-      return `css=*[${attr}="${value}"]`
-    }
-  }
-  return null
-})
-
-LocatorBuilders.add('id', function id(e) {
-  if (e.id && (e.id.split(/\d+/).length == 1 || this.appType != 'MN')) {
-    return 'id=' + e.id
-  }
-  return null
-})
-
-LocatorBuilders.add('linkText', function linkText(e) {
-  if (e.nodeName == 'A') {
-    let text = e.textContent
-    if (!text.match(/^\s*$/)) {
-      return (
-        'linkText=' + text.replace(/\xA0/g, ' ').replace(/^\s*(.*?)\s*$/, '$1')
-      )
-    }
-  }
-  return null
-})
-
-LocatorBuilders.add('name', function name(e) {
-  if (e.name) {
-    if (this.appType != 'MN')
-      return 'name=' + e.name
-    var splitByNumber = e.name.split(/\d+/)
-    if (splitByNumber.length == 1) {
-      return 'name=' + e.name
-    } else if (splitByNumber.length > 1) {
-      return 'xpath=//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[' + this.getXpathForAttribute('name', e.name) + ']'
-    }
-  }
-  return null
-})
-
-/*LocatorBuilders.add('css:finder', function cssFinder(e) {
-  return 'css=' + finder(e)
-})*/
-
-LocatorBuilders.add('xpath:link', function xpathLink(e) {
-  if (e.nodeName == 'A') {
-    let text = e.textContent
-    if (!text.match(/^\s*$/)) {
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('a') +
-          "[contains(text(),'" +
-          text.replace(/^\s+/, '').replace(/\s+$/, '') +
-          "')]",
-        e
-      )
-    }
-  }
-  return null
-})
-
-LocatorBuilders.add('xpath:img', function xpathImg(e) {
-  if (e.nodeName == 'IMG') {
-    if (e.alt != '') {
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('img') +
-          '[@alt=' +
-          this.attributeValue(e.alt) +
-          ']',
-        e
-      )
-    } else if (e.title != '') {
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('img') +
-          '[@title=' +
-          this.attributeValue(e.title) +
-          ']',
-        e
-      )
-    } else if (e.src != '') {
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('img') +
-          '[contains(@src,' +
-          this.attributeValue(e.src) +
-          ')]',
-        e
-      )
-    }
-  }
-  return null
-})
-
-LocatorBuilders.add('xpath:attributes', function xpathAttr(e) {
-  let PREFERRED_ATTRIBUTES = [
-    'id',
-    'name',
-    'value',
-    'type',
-    'action',
-    'onclick',
-  ]
-  if (this.appType == 'MN') {
-    PREFERRED_ATTRIBUTES = [
-      'comppath',
-      'name',
-      'value',
-      'type',
-      'action',
-      'onclick'
-    ]
-  } else if (this.appType == 'Flex') {
-    PREFERRED_ATTRIBUTES = [
-      'name',
-      'id',
-      'title',
-      'value',
-      'type',
-      'action',
-      'onclick'
-    ]
-  }
+LocatorBuilders.prototype.xpathAttr = function(e) {
   let i = 0
-
+  var self = this;
   function attributesXPath(name, attNames, attributes) {
     let locator = '//' + this.xpathHtmlElement(name) + '['
     for (i = 0; i < attNames.length; i++) {
@@ -431,7 +343,7 @@ LocatorBuilders.add('xpath:attributes', function xpathAttr(e) {
       if (attName == 'action' || attName == 'value' || attName == 'onclick')
         locator += '@' + attName + '=' + this.attributeValue(attributes[attName])
       else
-        locator += this.getXpathForAttribute(attName, attributes[attName])
+        locator += self.getXpathForAttribute(attName, attributes[attName])
     }
     locator += ']'
     return this.preciseXPath(locator, e)
@@ -446,8 +358,8 @@ LocatorBuilders.add('xpath:attributes', function xpathAttr(e) {
     }
     let names = []
     // try preferred attributes
-    for (i = 0; i < PREFERRED_ATTRIBUTES.length; i++) {
-      let name = PREFERRED_ATTRIBUTES[i]
+    for (i = 0; i < LocatorBuilders.PREFERRED_ATTRIBUTES.length; i++) {
+      let name = LocatorBuilders.PREFERRED_ATTRIBUTES[i]
       if (attsMap[name] != null) {
         names.push(name)
         let locator = attributesXPath.call(
@@ -463,190 +375,6 @@ LocatorBuilders.add('xpath:attributes', function xpathAttr(e) {
     }
   }
   return null
-})
-
-LocatorBuilders.add('xpath:idRelative', function xpathIdRelative(e) {
-  let path = ''
-  let current = e
-  while (current != null) {
-    if (current.parentNode != null) {
-      path = this.relativeXPathFromParent(current) + path
-      if (
-        1 == current.parentNode.nodeType && // ELEMENT_NODE
-        current.parentNode.getAttribute('id')
-      ) {
-        return this.preciseXPath(
-          '//' +
-            this.xpathHtmlElement(current.parentNode.nodeName.toLowerCase()) +
-            '[@id=' +
-            this.attributeValue(current.parentNode.getAttribute('id')) +
-            ']' +
-            path,
-          e
-        )
-      }
-    } else {
-      return null
-    }
-    current = current.parentNode
-  }
-  return null
-})
-
-LocatorBuilders.add('xpath:href', function xpathHref(e) {
-  if (e.attributes && e.hasAttribute('href')) {
-    let href = e.getAttribute('href')
-    if (href.search(/^http?:\/\//) >= 0) {
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('a') +
-          '[@href=' +
-          this.attributeValue(href) +
-          ']',
-        e
-      )
-    } else {
-      // use contains(), because in IE getAttribute("href") will return absolute path
-      return this.preciseXPath(
-        '//' +
-          this.xpathHtmlElement('a') +
-          '[contains(@href, ' +
-          this.attributeValue(href) +
-          ')]',
-        e
-      )
-    }
-  }
-  return null
-})
-
-LocatorBuilders.add('xpath:position', function xpathPosition(
-  e,
-  opt_contextNode
-) {
-  let path = ''
-  let current = e
-  while (current != null && current != opt_contextNode) {
-    let currentPath
-    if (current.parentNode != null) {
-      currentPath = this.relativeXPathFromParent(current)
-    } else {
-      currentPath = '/' + this.xpathHtmlElement(current.nodeName.toLowerCase())
-    }
-    path = currentPath + path
-    let locator = '/' + path
-    if (e == this.findElement(locator)) {
-      return 'xpath=' + locator
-    }
-    current = current.parentNode
-  }
-  return null
-})
-
-LocatorBuilders.add('xpath:innerText', function xpathInnerText(el) {
-  if (el.innerText) {
-    return `xpath=//${el.nodeName.toLowerCase()}[contains(.,'${el.innerText}')]`
-  } else {
-    return null
-  }
-})
-
-LocatorBuilders.prototype.buildReactTableRowData = function(e) {
-  try {
-    if (this.appType == 'BOB') {
-      //Table check
-      var closestTable = e.closest('table[id*=-body]');
-      var tableInModal = false;
-      if (closestTable == null || closestTable == undefined) {
-        closestTable = e.closest('table[class*=slds-table_fixed-layout]');
-        tableInModal = true;
-      }
-      if (closestTable == null || closestTable == undefined)
-        return undefined;
-      var tableClass = closestTable.getAttribute('class')
-      if (tableClass == null || tableClass == undefined || (tableClass != undefined && !tableClass.toLowerCase().includes('slds-table')))
-        return undefined;
-      var tableRowData = {};
-      var elementType = '', innerText = '', columnName = '', columnType = ''
-      if (tableInModal) {
-        var tdEl = e
-        if (e.nodeName.toLowerCase() != 'td') {
-          tdEl = e.closest('td[role=gridcell]')
-        }
-        var finalEl = tdEl.querySelectorAll('[id*=-content]')[0]
-        if (finalEl == undefined || finalEl == null) {
-          finalEl = tdEl.querySelectorAll('button')
-          elementType = 'BUTTON'
-          columnType = 'BLANK_INDEX'
-        } else {
-          innerText = finalEl.textContent.trim()
-          var tdIdx = Array.prototype.slice.call(tdEl.parentNode.children).indexOf(tdEl)
-          var theadRow = closestTable.querySelectorAll('thead')[0].children[0]
-          var reqTableHead = theadRow.children[tdIdx]
-          columnName = reqTableHead.querySelectorAll('span[id*=-text][class*=slds-truncate]')[0].textContent.trim()
-          elementType = 'PLAIN_TEXT'
-          columnType = 'HEADER_NAME'
-        }
-      } else {
-        //if event in td element itself && not stacked
-        if (e.nodeName.toLowerCase() == 'td' && e.children.length != 1)
-          return undefined;
-        var divEl = e.closest('div[class=slds-truncate]')
-        if (e.nodeName.toLowerCase() == 'td' && (divEl == null || divEl == undefined))
-          divEl = e.querySelectorAll('div[class=slds-truncate]')[0]
-        if (divEl == null || divEl == undefined) {
-          var tdEl = e.closest('td[role=gridcell]')
-          if (tdEl) {
-            var checkBoxEl = tdEl.querySelectorAll('input[type=checkbox]')
-            var buttonEl = tdEl.querySelectorAll('button')
-            if (checkBoxEl && checkBoxEl[0]) {
-              elementType = 'OTHER'
-              columnType = 'CHECKBOX_INDEX'
-              var isChecked = checkBoxEl[0].getAttribute('checked') || checkBoxEl[0].getAttribute('checked') == 'true'
-              innerText = isChecked ? 'on' : 'off';
-            }
-          }
-        } else {
-          columnType = 'HEADER_NAME'
-          var tdEl = divEl.parentNode;
-          var tdIdx = Array.prototype.slice.call(tdEl.parentNode.children).indexOf(tdEl)
-          var divIdx = Array.prototype.slice.call(divEl.parentNode.children).indexOf(divEl)
-          var finalEl = divEl.querySelectorAll('[id*=-content]')[0];
-          if (!finalEl || finalEl == null) {
-            finalEl = divEl.querySelectorAll('div');
-            finalEl = finalEl[finalEl.length - 1]
-          }
-          var nodeName = finalEl.nodeName.toLowerCase();
-          if (finalEl.querySelectorAll('button').length > 0)
-            elementType = 'BUTTON'
-          else if (finalEl.querySelectorAll('use').length > 0)
-            elementType = 'IMAGE'
-          else if (nodeName == 'a') {
-            elementType = 'LINK'
-            innerText = finalEl.textContent.trim()
-          } else if ((nodeName == 'input' && eType != 'hidden') || nodeName == 'select')
-            elementType = 'OTHER'
-          else {
-            elementType = 'PLAIN_TEXT'
-            innerText = finalEl.textContent.trim()
-          }
-          var theadRow = closestTable.querySelectorAll('thead[class=data_table_head]')[0].children[0]
-          var reqTableHead = theadRow.children[tdIdx]
-          var reqDiv = reqTableHead.querySelectorAll('div[id*=-body]')[divIdx]
-          columnName = reqDiv.querySelectorAll('[class=slds-truncate]')[0].textContent.trim()
-        }
-      }
-      tableRowData.elementType = elementType
-      tableRowData.columnType = columnType
-      tableRowData.expectedValue = innerText
-      tableRowData.columnName = columnName
-      console.log(JSON.stringify(tableRowData))
-      return tableRowData;
-    }
-  } catch(e) {
-    console.error(e)
-  }
-  return undefined;
 }
 
 LocatorBuilders.prototype.getDisplayName = function(e, ignoreInnerText) {
@@ -673,57 +401,11 @@ LocatorBuilders.prototype.getDisplayName = function(e, ignoreInnerText) {
       }
     }
   }
-  if (this.appType == 'MN') {
-    var elXpath = this.getXpathOfAnElement(e, false);
-    if (elXpath) {
-      var labelXpath = elXpath + '/preceding::*[((local-name() = \'span\' and contains(@domattr,\'extField\')) or (local-name() = \'label\' and contains(@class,\'left\'))) and string-length(normalize-space(text())) > 1][1]';
-      var labelEl = this.findElement(labelXpath);
-      if (labelEl && labelEl.innerText) {
-        var innerText = this.getInnerTextWithoutChildren(labelEl);
-        if (innerText.length > 1)
-          return innerText;
-      }
-    }
-  }
   return undefined;
 }
 
 LocatorBuilders.prototype.logging = function(message) {
     browser.runtime.sendMessage({type: "bglog", payload: message});
-}
-
-LocatorBuilders.prototype.setPreferredOrderByAppType = function(appType) {
-  this.appType = appType;
-  if (appType == 'MN')
-    LocatorBuilders.setPreferredOrder('leftNav,table,xpath:comppath,xpath:popupsinmn,xpath:comppathRelative,xpath:attributes,xpath:link,linkText,name,' +
-        'xpath:innerText,xpath:img,id,xpath:idRelative,xpath:href,xpath:position,css:data-attr');
-  else if (appType == 'Flex')
-    LocatorBuilders.setPreferredOrder('table,xpath:attributes,name,id,xpath:link,xpath:idRelative,xpath:innerText,linkText,' +
-        'xpath:img,xpath:href,xpath:position,css:data-attr,leftNav,xpath:comppath,xpath:comppathRelative,xpath:popupsinmn');
-  else if (appType == 'BOB')
-    LocatorBuilders.setPreferredOrder('table,id,linkText,name,css:data-attr,xpath:link,xpath:img,xpath:idRelative,xpath:attributes,' +
-        'xpath:href,xpath:position,xpath:innerText,leftNav,xpath:comppath,xpath:comppathRelative,xpath:popupsinmn');
-  else
-    LocatorBuilders.setPreferredOrder('id,linkText,name,css:data-attr,xpath:link,xpath:img,xpath:attributes,xpath:idRelative,' +
-        'xpath:href,xpath:position,xpath:innerText,table,leftNav,xpath:comppath,xpath:comppathRelative,xpath:popupsinmn');
-  console.log('Updated order for App Type: ' + this.appType);
-}
-
-LocatorBuilders.prototype.setPreferredOrderByAppTypeFirstTime = function() {
-  var self = this;
-  function getAppType() {
-    let appType = undefined;
-    browser.runtime.sendMessage({type: "getAppType"}).then(function(response) {
-      if (response && response.appType) {
-        appType = response.appType;
-        self.setPreferredOrderByAppType(appType);
-        console.log('Updated First time order for App Type: ' + self.appType);
-      }
-    });
-    return appType;
-  }
-  let appType = getAppType();
-  self.setPreferredOrderByAppType(appType);
 }
 
 LocatorBuilders.prototype.isElementUniqueWithXPath = function(xpath, e) {
@@ -755,11 +437,6 @@ LocatorBuilders.prototype.getXpathForAttribute = function(attrName, attrValue) {
   return xpathCond;
 }
 
-LocatorBuilders.prototype.getXpathFromComppathAttr = function(nodeName, comppath) {
-    var xpathCond = '[' + this.getXpathForAttribute('comppath', comppath) + ']';
-    return '//' + this.xpathHtmlElement(nodeName.toLowerCase()) + xpathCond;
-}
-
 LocatorBuilders.prototype.getInnerTextWithoutChildren = function(node) {
   let child = node.firstChild, texts = [];
   while (child) {
@@ -770,276 +447,3 @@ LocatorBuilders.prototype.getInnerTextWithoutChildren = function(node) {
   }
   return texts.join('').trim()
 }
-
-LocatorBuilders.prototype.getXpathOfAnElement = function(e, skipCases) {
-  var elXpath = undefined;
-  if (this.appType == 'MN' || this.appType == 'Flex' || this.appType == "BOB") {
-    if (this.appType == 'MN') {
-      var comppath = e.getAttribute('comppath');
-      if (comppath) {
-        if (comppath.indexOf('spreadsheetContainer') > -1 && skipCases)
-          return undefined;
-        elXpath = this.preciseXPath('//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[@comppath=' + this.attributeValue(comppath) + ']', e);
-        elXpath = elXpath.substring(6);
-        return elXpath;
-      }
-    }
-    if (this.appType == 'Flex' && e.getAttribute('title')) {
-      elXpath = '//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[@title=' + this.attributeValue(e.getAttribute('title')) + ']';
-      if (e == this.findElement('xpath=' + elXpath)) {
-        return elXpath;
-      }
-    }
-    if (e.id) {
-      elXpath = '//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[@id=' + this.attributeValue(e.id) + ']';
-      if (e == this.findElement('id=' + e.id)) {
-        return elXpath;
-      }
-    }
-    if (e.name) {
-      elXpath = '//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[@name=' + this.attributeValue(e.name) + ']';
-      if (e == this.findElement('name=' + e.name)) {
-        return elXpath;
-      } else {
-        elXpath = this.preciseXPath(elXpath, e);
-        elXpath = elXpath.substring(6);
-        return elXpath;
-      }
-    }
-  }
-  return undefined;
-}
-
-LocatorBuilders.add('leftNav', function leftNav(e) {
-  if (this.appType == 'MN') {
-    var current = e;
-    if (current.parentNode != null) {
-      var parentNode = current.parentNode;
-      var isLeftNav = (parentNode.nodeName.toLowerCase() == 'div' ? (parentNode.hasAttribute('open') && parentNode.hasAttribute('isleaf')) : false);
-      if ( isLeftNav ) {
-        var displayName = undefined;
-        var nodeName = current.nodeName.toLowerCase();
-        var isToggleBtn = nodeName == 'span' && current.hasAttribute('class') && current.getAttribute('class').includes('clickable');
-        var elXpath = this.xpathHtmlElement(nodeName);
-        var elInnerText;
-        if (isToggleBtn) {
-          elXpath = elXpath + '[contains(@class,\'clickable\')]';
-          var nextEl = e.nextSibling;
-          elInnerText = this.getInnerTextWithoutChildren(nextEl);
-          displayName = 'expand/collapse button next to ' + elInnerText;
-        } else if (nodeName == 'a') {
-          elInnerText = this.getInnerTextWithoutChildren(current);
-          elXpath = elXpath + '[contains(.,\'' + elInnerText + '\')]';
-        } else {
-          return null;
-        }
-        var links = '',dynamicLinks = [];
-        while (parentNode != null && isLeftNav) {
-          var isLeaf = parentNode.hasAttribute('isleaf') ? parentNode.getAttribute('isleaf') : undefined;
-          var curText = this.getInnerTextWithoutChildren(parentNode.querySelectorAll('a')[0]);
-          links = curText + '->' + links;
-          if (isLeaf == 'false')
-            dynamicLinks.push(curText);
-          var nameAttr = parentNode.hasAttribute('name') ? ('-node-' + parentNode.getAttribute('name').split('-node-')[1]) : undefined;
-          var curElXpath = 'div[';
-          if (nameAttr)
-            curElXpath = curElXpath + 'contains(@name,\'' + nameAttr + '\')';
-          else if (elInnerText)
-            curElXpath = curElXpath + 'contains(.,\'' + elInnerText + '\')';
-          else
-            return null;
-          elXpath = curElXpath + ']/' + elXpath;
-          parentNode = parentNode.parentNode;
-          if (parentNode != null) {
-            isLeftNav = (parentNode.nodeName.toLowerCase() == 'div' ? (parentNode.hasAttribute('open') && parentNode.hasAttribute('isleaf')) : false);
-          }
-        }
-        if (this.isElementUniqueWithXPath('//' + elXpath, e)) {
-          if (dynamicLinks.length >= 2 && !this.recordedType) {
-            this.recordedType = 'leftNav';
-            var navLinks = '';
-            //Ignore last element
-            for (var k = 0 ; k < dynamicLinks.length - 1; k++) {
-              navLinks = dynamicLinks[k] + (navLinks == '' ? '' : '->' + navLinks);
-            }
-            this.additionalData = 'navLinks=' + navLinks;
-            this.logging('From leftNav: ' + this.additionalData);
-          }
-          if (displayName && !this.displayName)
-            this.displayName = displayName;
-          return "xpath=//" + elXpath;
-        }
-      }
-    }
-  }
-  return null;
-})
-
-LocatorBuilders.add('table', function table(e) {
-  if (this.appType == 'BOB') {
-    var tableRowData = this.buildReactTableRowData(e);
-    if (tableRowData) {
-      if (!this.recordedType) {
-        this.recordedType = 'table';
-        this.additionalData = 'rowType=data|elementType=' + tableRowData.elementType + '|columnName=' + tableRowData.columnName + '|columnType=' + tableRowData.columnType;
-        this.displayName = undefined;
-      }
-      var id = (tableRowData.columnName && tableRowData.columnName != '' ? tableRowData.columnName : (tableRowData.columnType != '' ? tableRowData.columnType : undefined))
-      if (!id)
-        id = e.id ? e.id : (e.name ? e.name : e.getAttribute('data-id'));
-      return 'table=' + id;
-    }
-    return undefined;
-  }
-  var elXpath = this.getXpathOfAnElement(e, true);
-  if (elXpath) {
-    var parXpath = '';
-    if (this.appType == 'MN') {
-      parXpath = elXpath + '/ancestor::tr[contains(@class,\'tableRow bodyRow bodyRow-\') or contains(@class,\'tableRow headerRow headerRow-\')]';
-      // or contains(@class,\'tableRow footerRow footerRow-\')
-    } else {
-      parXpath = elXpath + '/ancestor::div[contains(@id,\'tbl-container\')]'
-    }
-    var tdEl = this.findElement(parXpath);
-    if (tdEl) {
-      var rowType = 'data';
-      if (tdEl.getAttribute('class').includes('headerRow-')) {
-        rowType = 'header';
-      }/*else if (tdEl.getAttribute('class').includes('footerRow-')) {
-        rowType = 'footer';
-      }*/
-      var elementType = '';
-      var nodeName = e.nodeName.toLowerCase();
-      var eType = e.type ? e.type.toLowerCase() : '';
-      var classes = e.getAttribute('class') ? e.getAttribute('class').toLowerCase() : '';
-      if (nodeName == 'span' && !classes.includes('anchor') && !classes.includes('fa-fw'))
-        elementType = 'PLAIN_TEXT'
-      else if (nodeName == 'a' || (nodeName == 'span' && classes.includes('anchor')))
-        elementType = 'LINK'
-      else if ((nodeName == 'input' && eType != 'hidden') || nodeName == 'select')
-        elementType = 'OTHER'
-      else if (nodeName == 'button')
-        elementType = 'BUTTON'
-      else if (nodeName == 'img')
-        elementType = 'IMAGE'
-      if (!this.recordedType) {
-        this.recordedType = 'table';
-        this.additionalData = 'rowType=' + rowType + '|elementType=' + elementType;
-        this.displayName = undefined;
-      }
-      return 'table=' + elXpath;
-    }
-  }
-  return null;
-})
-
-LocatorBuilders.add('xpath:comppath', function xpathComppath(e) {
-    var comppath = e.getAttribute('comppath');
-    if (comppath) {
-    let locator = this.getXpathFromComppathAttr(e.nodeName, comppath);
-    //If comppath matches exactly with element found then only return comppath
-    if (this.isElementUniqueWithXPath(locator, e))
-      return 'xpath=' + locator;
-    else {
-      if (e.innerText) {
-        let textLoc = locator + '[contains(.,\'' + this.getInnerTextWithoutChildren(e) + '\')]';
-        if (this.isElementUniqueWithXPath(textLoc, e))
-          return 'xpath=' + textLoc;
-      }
-      if (comppath.indexOf('spreadsheetContainer') > -1) {
-        let labelLoc = '//' + this.xpathHtmlElement(e.nodeName.toLowerCase()) + '[@comppath=' + this.attributeValue(comppath) + ']' +
-            '/ancestor::tr[contains(@class, \'tableRow bodyRow\')]/td[1]/span';
-        let labelEl = this.findElement(labelLoc);
-        if (labelEl && labelEl.innerText) {
-          let finalLoc = '//td[(normalize-space(.)=\'' + this.getInnerTextWithoutChildren(labelEl) + '\')]//following::' + locator.substring(2);
-          if (comppath.indexOf('spreadsheetContainer.FormularyCondition') > -1) {
-            let labelParLoc = labelLoc + '/ancestor::div[contains(@comppath,\'spreadsheetContainer.FormularyCondition\')]//div[contains(@class,\'title\')]';
-            let labelParEl = this.findElement(labelParLoc);
-            if (labelParEl && labelParEl.innerText)
-              finalLoc = '//div[(translate(normalize-space(.),\'ABCDEFGHIJKLMNOPQRSTUVWXYZ\',\'abcdefghijklmnopqrstuvwxyz\')=\'' + this.getInnerTextWithoutChildren(labelParEl).toLowerCase() + '\')]' +
-                  '//following::' + finalLoc.substring(2);
-          }
-          //this.logging('Final Locator: ' + finalLoc);
-          if (e == this.findElement(finalLoc))
-            return 'xpath=' + finalLoc;
-        }
-      }
-    }
-  }
-  return null;
-})
-
-LocatorBuilders.add('xpath:comppathRelative', function xpathComppathRelative(e) {
-    let path = '';
-    let current = e;
-    while (current != null) {
-      if (current.parentNode != null) {
-        path = this.relativeXPathFromParent(current) + path;
-        if (
-            1 == current.parentNode.nodeType && // ELEMENT_NODE
-            current.parentNode.getAttribute('comppath')
-        ) {
-            let locator = this.getXpathFromComppathAttr(current.parentNode.nodeName, current.parentNode.getAttribute('comppath'));
-            //this.logging('inside xpath:comppathRelative: ' + locator);
-            if (this.isElementUniqueWithXPath(locator, current.parentNode)) {
-              if (current.parentNode.nodeName.toLowerCase() == 'a') {
-                this.displayName = (current.parentNode.innerText && current.parentNode.innerText.trim().length > 1) ? current.parentNode.innerText.trim() :
-                    (this.displayName ? this.displayName : this.getDisplayName(current.parentNode));
-                return 'xpath=' + locator;
-              }
-              return 'xpath=' + locator + path;
-            }
-        }
-      } else {
-          return null;
-      }
-    current = current.parentNode;
-  }
-  return null;
-})
-
-LocatorBuilders.add('xpath:popupsinmn', function specialCasesInMN(e) {
-  if (this.appType != 'MN')
-    return null;
-  var detailTitleBar = e.closest('[class*=detailTitleBar]');
-  var alertBar = e.closest('[class*=showAlert]');
-  var msgPopup = e.closest('[class*=CMnDisplayMsgComp],[class*=CMnPopupComp]');
-  var errorIcon = e.closest('[class*=fieldError]');// span class=fieldError for error icon
-  var errorMsgPopup = e.closest('[class*=validationContainer]');
-  var locator;
-  if (detailTitleBar || alertBar) {
-    //find parent element for which comppath is defined and build locator relative to that
-    var comppathEl = e.closest('[comppath*=root]');
-    if (comppathEl) {
-      var comppath = comppathEl.getAttribute('comppath');
-      var titleClass = detailTitleBar ? 'detailTitleBar' : 'showAlert';
-      locator = "//*[@comppath='" + comppath + "']" + "//*[contains(@class,'" + titleClass + "')]//*[normalize-space(@class)='title']";
-    }
-  } else if (msgPopup) {
-    if (e.closest('table')) {
-      locator = "//*[contains(@class,'CMnDisplayMsgComp') or contains(@class,'CMnPopupComp')]//*[normalize-space(@class)='msgText']";
-    }
-  } else if (errorMsgPopup) {
-    locator = "//*[contains(@class,'validationContainer')]";
-    var elClass = e.getAttribute('class');
-    var elStyle = e.getAttribute('style');
-    //message is present at 2nd div child of above validationContainer
-    if (elClass && elClass.indexOf('closeLink') > -1)
-      locator = locator + "//*[contains(@class,'closeLink')]";
-    else if (elStyle && elStyle.indexOf('block') > -1)
-      locator = locator + "/div[2]";
-  } else if (errorIcon) {
-    //Below may need special handling if more than one error field appears at once on the page
-    //currently handled for one error field case only
-    locator = "//*[contains(@class,'fieldError') and contains(@class,'fa-fw')]";
-    if (this.isElementUniqueWithXPath(locator, e))
-      return 'xpath=' + locator;
-    else
-      locator = undefined;
-  }
-  if (locator)
-    return "xpath=" + locator;
-  return null;
-})
-
-LocatorBuilders.prototype.setPreferredOrderByAppTypeFirstTime()
